test: isolate SSE progress test
