+++ backend/tests/additionalApi.test.js
@@ -401,15 +401,6 @@ test('POST /api/admin/competitions unauthorized', async () => {
   expect(res.status).toBe(401);
 });
 
-test('SSE progress endpoint streams updates', async () => {
-  const req = request(app).get('/api/progress/job1');
-  setTimeout(() => {
-    progressEmitter.emit('progress', { jobId: 'job1', progress: 100 });
-  }, 50);
-  const res = await req;
-  expect(res.text).toContain('data: {"jobId":"job1","progress":100}');
-});
-
 test('POST /api/create-order rejects unknown job', async () => {
   db.query.mockResolvedValueOnce({ rows: [] });
   const res = await request(app).post('/api/create-order').send({ jobId: 'bad' });
